Updated README.md to have merge as default option for pull (#7)
## Summary

(Summarize the story / bug encountered concisely, link to the story /
bug )

## What this PR do?

(What this PR aim to achieve? what changes it brings to user or
customer? For example, If this is API change please mention curl
request/postman request/open API to describe the change)

## Where should the reviewer start?

(Where the reviewer should start. Please guide the reviewer to how to go
through the CL to review it more effectively)

## Recommendations on how to test or use this change

(How the changes are tested and how other can use and test it)

## Link to related and dependent issues

(Please mention related issues)

## Does anything need to be done? (ex. Notify someone, deploy sth else)
before deploying this change?

(Please mention open points or issues or remaining things)